Configurable VMAR allocator
