+++ kernel/Cargo.toml
@@ -66,9 +66,11 @@ fixed = "1.28.0"
 riscv = { version = "0.11.1", features = ["s-mode"] }
 
 [features]
-all = ["cvm_guest"]
+default = ["dist_vmar_alloc"]
+all = ["cvm_guest", "dist_vmar_alloc"]
 
 cvm_guest = ["dep:tdx-guest", "ostd/cvm_guest"]
+dist_vmar_alloc = []
 
 [lints]
 workspace = true
